+++ slack.py
@@ -1,0 +1,40 @@
+import logging
+
+import requests
+
+
+class SlackUpdater(object):
+    def __init__(self, channel, token, bot_name):
+        logging.basicConfig(level=logging.INFO)
+        self.channel = channel
+        self.token = token
+        self.bot_name = bot_name
+
+    def post_slack_message(self, payload):
+        requests.post('https://slack.com/api/chat.postMessage',
+                      data={
+                          'channel': self.channel,
+                          'token': self.token,
+                          'text': payload,
+                          'username': self.bot_name
+                      })
+
+        logging.info("\nPosting to Slack: done")
+
+    def prepare_slack_update(self, tickets):
+        """
+        Processes acquired results
+        """
+        result = ''
+        for release, tickets_list in tickets.iteritems():
+            result += '*' + release + '*\n'
+
+            if len(tickets_list) == 0:
+                result += 'Nothing user facing\n'
+            else:
+                result += '```'
+                for ticket in tickets_list:
+                    result += 'https://wikia-inc.atlassian.net/browse/' + ticket['key'] + ' ' + ticket['desc'] + '\n'
+                result += '```\n'
+            result += '\n'
+        return result
